First successful MPI execution using point-to-point communication
This version has a basic point-to-point communication based MPI amongst
multiple processes. A root, or master, process handles the initial read of
the data to get the total number of rows in the text file

This data point is used to distribute read indices amongst the workers.
Each worker receives multiple tuples of start/end read indices to process.
The workers read in chunks, but receives indices that is all emcompassing
of the file

This was only tested on a local node (desktop computer) but shows dramatic
speedups, taking in total 2 - 3 minutes to process the data-big.txt file