Fixed major error. Now it does not sqrt X^2 + Y^2, comparing to R^2
Originally, the code was squaring the radius, then comparing that to the
sqrt of X^2 + Y^2. This was a mistake, and may have affected the Pi
estimation

The next commit will run the code again for 500 million to 1 billion
simulation points to see if it converges to a more acceptable value. In
the old code, even with 100 - 500 million points, decimal precision was no
better than 2 decimal places