Newline character removal optimization added to scrub.py
Previously, newlines \n and \r\n was first detected in first and last row.
If it was found, use string split method to split by the found newline
character, keeping only the string before the newline. This was known to
be slow, and a new method is now implemented that achieves the same result
but faster

This method is based on string slicing. When \n or \r\n is detected, the
last 1 or 2 characters (excluding backslash) is removed using :-1 or :-2
negative indexing. Using iPython's timeit magic function, this was tested
on data.txt file with 1000 runs, picking the best of each. On average, the
negative indexing method was roughly 2 - 3 times faster than string split

For data-big.txt, the final different was small however, with roughly 200
- 1000 lines parsed extra per second. Total time decrease ranged from 0.1
  to 0.5 seconds faster. For larger datasets however, this can very
  quickly add up