File reading optimized for avoiding re-reading whole to get to line
A new class definition Rowread is developed to replace row_reader
generator. Previously, row_reader required the file, start and end line
numbers to read the appopriate file section for each process

Each worker process was provided a set of start/end row line number
(index) to read. Every time the process finishes a start/end row line
number, it had to re-read the file from the start when it started work on
the next start/end row line number. This was slow and inefficient as it
had to skip, from the start, all rows until it hit the start line number

To avoid this, a new class is defined, Rowread, that internally maintains an
open file connection. Using open, it remembers the last row position it
finished at from before. When it starts the next chunk, it won't have to
start from the beginning of the file to get to the next start line number.
This optimization requires that each subsequent start/end row line numbers
is further down the file than where it previously stopped

To further the effiency of this class, the start/end row line numbers for
any given progress is right next to each other. For example, first batch
for a process is 0 - 1000. Then the next batch is 1001 - 2000, etc. This
flanked indices ensure minimized file seeking

The reading optimization provides a sizeable speed up of approximately
60000 rows parsed per second, bringing the speed of the program to about
560000 rows parsed per second in total. This is brings the current
program, with its previous optimizations, to about 2.07 times faster than
the program without these optimizations