Implemented parallelized Jarque-Bera normality test in normal.py
The Jarque-Bera test is based off of Normal Distribution's skewness and
kurtosis. If the series' skewness and kurtosis is different from a normal
distribution, the test states that we can reject the null (series is
normal)

The parallelization comes from the fact that standard deviation, skewness,
and kurtosis have summable parts. The numerators and denominators are
computed in parts amongst the workers. The results are sent to the root
process and the Jarque-Bera test statistic is computed and compared

The wrinkle was the fact that standard deviation, skewness, and kurtosis
all required the series' sample mean. Therefore, the sample mean is first
computed amongst the workers (using a similar methodology), aggregated at
the root process, then sent back to the workers for computing the partial
standard deviations, skewness, and kurtosis