Switch to Python 2.7, slight regex change but see massive performance gain
The program has been tested with Python 2.7, which required a slight row
regex search to ensure that it works

This change however saw a massive speed increase. Previously, running on
Python 2.6, scrub program was iterating for about 520k - 560k rows
per second. When switched to Python 2.7, this has increased to about 1.1
million rows per second. Compared to the older version on Python 2.6,
which averaged roughly 270k rows per second, the optimizations and Python
2.7 gave us a 4x speed increase

Other smaller changes include the regex alteration. The old style didn't
work with Python 2.7, which had to do with the way newline was handled in
the different Python versions (maybe, its a guess). But also changes to
the ResultLogger API, now requiring KV items to be provided in bulk
instead of one at a time. This was done to ensure the separator lined up
properly in the output